creating a lot of things :)

+++ controllers/admin/ArticleController.php
@@ -5,10 +5,32 @@ use app\models\admin\UpdateForm;
 use app\models\Article;
 use app\models\Category;
 use app\models\Company;
+use app\models\User;
+use yii\db\Exception;
+use yii\filters\AccessControl;
 use yii\web\Controller;
+use yii\web\NotFoundHttpException;
+use yii\web\Response;
+
 
 class ArticleController extends Controller
 {
+
+    public function behaviors()
+    {
+        return [
+            'access' => [
+                'class' => AccessControl::className(),
+                'rules' => [
+                    [
+                        'allow' => true,
+                        'roles' => ['@'],
+                    ],
+                ],
+            ],
+        ];
+    }
+
     public $layout = 'admin';
 
     public function actionIndex()
@@ -22,12 +44,11 @@ class ArticleController extends Controller
 
     public function actionCreate()
     {
-        $article = Article::find()->one();
-        $createdArticle = new Article();
+        $article = new Article();
 
-        if ($createdArticle->load(\Yii::$app->request->post())) {
-            if ($createdArticle->validate()) {
-                $createdArticle->save();
+        if ($article->load(\Yii::$app->request->post())) {
+            if ($article->validate()) {
+                $article->save();
 
                 return $this->redirect('index');
             }
@@ -37,6 +58,18 @@ class ArticleController extends Controller
             'article' => $article,
             'categories' => Category::find()->all(),
             'companies'  => Company::find()->all(),
+        ]);
+    }
+
+
+    public function actionView($id)
+    {
+        //?????????????????????????
+        $articles = Article::find()->joinWith(['category', 'company'])->all();
+
+        return $this->render('view', [
+            'articles' => $articles,
+            'id' => $id,
         ]);
     }
 
@@ -56,36 +89,39 @@ class ArticleController extends Controller
         return $this->render('update', [
             'article' => $article,
             'categories' => Category::find()->all(),
+            'companies' => Company::find()->all(),
         ]);
     }
 
 
-    public function actionDelete($id, $status = 0)
+    public function actionDelete($id)
     {
-        switch($status)
-        {
-            case 0:
-                return $this->render('delete', ['id' => $id]);
-                break;
-
-            case 1:
-                $article = Article::findOne(['id' => $id]);
-                $article->delete();
-                return $this->redirect('index');
-                break;
-
-            case 2:
-                return $this->redirect('index');
-                break;
+        $article = Article::findOne(['id' => $id]);
+        if ($article === null) {
+            if (\Yii::$app->request->isAjax) {
+                \Yii::$app->response->format = Response::FORMAT_JSON;
+            }
+            throw new NotFoundHttpException('Current record not found.');
         }
 
+//        if (\Yii::$app->request->isAjax) {
+//            try {
+//                if (true) {
+//                if ($article->delete() !== false) {
+//                    \Yii::$app->response->format = Response::FORMAT_JSON;
+//                    return ['status' => 'ok'];
+//                }
+//            } catch (\Throwable $e) {
+//                // todo
+//            }
+//        }
+
+        if (\Yii::$app->request->isPost) {
+            $article->delete();
+            return $this->redirect('index');
+        }
+
+        return $this->render('delete', ['article' => $article]);
     }
-
-
-    public function actionView()
-    {
-
-    }
-
 
 }
